break out handlebars helpers into own file; remove commented out hook;
